+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add PostgreSQLAdapter#set_replica_identity
+
 ## [0.1.1] - 2021-06-07
 
 - Add PostgreSQLAdapter#defer_constraints